refactor: move `LightboxOverlay` alongside `TemplateRenderer`
