+++ packages/next/src/index.ts
@@ -8,4 +8,3 @@ export { Fonts } from './components/font';
 export { Image } from './components/image';
 export { Link } from './components/link';
 export { Script } from './components/script';
-export { LightboxOverlay } from './components/lightbox-overlay';
