+++ packages/next/src/template-renderer/lightbox-overlay.tsx
